backend/requirements.txt
# EduAI Backend - Production Requirements (Docker/Python 3.11)

# ── Web Framework ──────────────────────────────────────────────────
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.7
werkzeug==3.0.3

# ── Database ───────────────────────────────────────────────────────
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
alembic==1.13.1

# ── Server ─────────────────────────────────────────────────────────
gunicorn==22.0.0

# ── Auth & Security ────────────────────────────────────────────────
bcrypt==4.1.3
pyjwt==2.8.0

# ── AI / OpenAI ────────────────────────────────────────────────────
openai==1.30.1
httpx==0.27.0
tiktoken==0.7.0

# ── OCR & Image Processing ─────────────────────────────────────────
Pillow==10.3.0
pytesseract==0.3.10
pdf2image==1.17.0

# ── Utils ──────────────────────────────────────────────────────────
python-dotenv==1.0.1
marshmallow==3.21.3
python-dateutil==2.9.0

# ── Cache / Queue (optional, won't crash if Redis unavailable) ─────
redis==5.0.4
